% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run Shiny App demonstrating the package}
\usage{
runExample(...)
}
\arguments{
\item{...}{passed to \code{\link[shiny:runApp]{shiny::runApp()}}, e.g. \code{port}, \code{launch.browser}}
}
\description{
Run Shiny App demonstrating the package
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  bpbounds::runExample()
}
}
